Flow chart
Hey Mr.Brooks,

This is my flow chart for the mini-game. Also, I just went over the google doc and rubric to double-check and I just realized, I totally missed the "make a simple game" part...
I got so concentrated into making the game that I kept adding more and more to it, I doubt it counts as a "simple game" but hopefully it's not too over the top.
In the end, I had a lot of fun and hope to do more assignments like this in the future!